[Feature #160784176] Restructure the code
