+++ app/__init__.py
@@ -7,7 +7,9 @@ included to make app a package
 from flask import Flask
 from flask_restful import Api
 
-from app.api.v1.views import Orders, OrderSpecific
+from app.api.v1.views_orders import Orders, OrderSpecific
+
+
 
 from instance.config import app_config
